cleared clutter and added layout
