+++ app/components/Navbar.tsx
@@ -1,15 +1,16 @@
 import { RiMoonFoggyFill } from 'react-icons/ri'
 import { BsFillSunFill } from 'react-icons/bs'
 import { animated } from 'react-spring'
-import { transition, useFadeIn, useExpand } from '~/theme/animate'
+import { transition, useFadeIn } from '~/theme/animate'
+import { useContext } from 'react'
+import { ThemeContext } from '~/theme/context'
 
 type Props = {
-  toggleTheme: () => void,
-  colorMode: 'dark' | 'light',
-  children: React.ReactNode,
+  // children: React.ReactNode,
 }
 
-const Navbar = ({ toggleTheme, colorMode, children }: Props) => {
+const Navbar = (props: Props) => {
+  const colorMode = useContext(ThemeContext)
 
   return (
     <>
@@ -24,13 +25,13 @@ const Navbar = ({ toggleTheme, colorMode, children }: Props) => {
           </h1>
           <button
             className={`p-2 mr-8 col-span-2 rounded-full shadow-xl mx-5 bg-amber-50 dark:bg-slate-800
-              border border-amber-200 dark:border-slate-800 ${colorMode === 'light' ? 'light-mode-shadow' : 'dark-mode-shadow'}
+              border border-amber-200 dark:border-slate-800 ${colorMode.theme === 'light' ? 'light-mode-shadow' : 'dark-mode-shadow'}
               hover:px-10 hover:mr-0 hover:border hover:rounded-full hover:border-amber-300  dark:hover:border-slate-600
               ${transition}`}
-            onClick={toggleTheme}
+            onClick={colorMode.toggleTheme}
           >
             {
-              colorMode === "light" ? (
+              colorMode.theme === "light" ? (
                 <BsFillSunFill size={32} className='text-amber-600' />
               ) : (
                 <RiMoonFoggyFill size={32} className='text-slate-200' />
@@ -39,17 +40,6 @@ const Navbar = ({ toggleTheme, colorMode, children }: Props) => {
           </button>
         </animated.div>
       </nav>
-      <div className=''>
-        {children}
-      </div>
-
-      <animated.div
-        style={useExpand()}
-        className={`text-center font-display py-5 mt-5 text-gray-700 dark:text-gray-400 drop-shadow-2xl
-         backdrop-blur-md ${transition}`}
-      >
-        Created by, Richie Varghese © 2022
-      </animated.div>
     </>
   )
 }
