+++ app/root.tsx
@@ -9,6 +9,8 @@ import {
 } from "@remix-run/react";
 import styles from "./tailwind.css"
 import appStyle from "./styles/app.css"
+import { useState, useEffect } from "react";
+import { ThemeContext } from "./theme/context";
 
 export const meta: MetaFunction = () => ({
   charset: "utf-8",
@@ -21,6 +23,19 @@ export function links() {
 }
 
 export default function App() {
+  const [colorMode, setColorMode] = useState<'dark' | 'light'>('light');
+  const toggleTheme = () => {
+    setColorMode((prevMode) => (prevMode === "light" ? "dark" : "light"));
+    localStorage.setItem("theme", colorMode === "light" ? "dark" : "light");
+  };
+
+  // Check if the user has a preference for dark mode, if so, set the theme to dark
+  useEffect(() => {
+    const localTheme = localStorage.getItem("theme");
+    localTheme && setColorMode(localTheme as 'dark' | 'light');
+  }, []);
+
+
   return (
     <html lang="en">
       <head>
@@ -28,10 +43,14 @@ export default function App() {
         <Links />
       </head>
       <body>
-        <Outlet />
-        <ScrollRestoration />
-        <LiveReload />
-        <Scripts />
+        <ThemeContext.Provider value={{ theme: colorMode, toggleTheme: toggleTheme }}>
+          <main className={`${colorMode}`} >
+            <Outlet />
+            <ScrollRestoration />
+            <LiveReload />
+            <Scripts />
+          </main>
+        </ThemeContext.Provider>
       </body>
     </html>
   );
