+++ app/theme/context.ts
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 
 export const ThemeContext = createContext({
-  theme: 'dark',
+  theme: 'dark' as 'dark' | 'light',
   toggleTheme: () => { },
 });
